# api/swagger.yaml
basePath: /
consumes:
    - application/json
definitions:
...
    AuthConfig:
        properties:
            password:
                type: string
                x-go-name: Password
            username:
                type: string
                x-go-name: Username
        type: object
        x-go-package: github.com/go-gost/x/config
...
    ChainConfig:
        properties:
            hops:
                items:
                    $ref: '#/definitions/HopConfig'
                type: array
                x-go-name: Hops
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
        type: object
        x-go-package: github.com/go-gost/x/config
...
    ConnectorConfig:
        properties:
            auth:
                $ref: '#/definitions/AuthConfig'
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            tls:
                $ref: '#/definitions/TLSConfig'
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
    DialerConfig:
        properties:
            auth:
                $ref: '#/definitions/AuthConfig'
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            tls:
                $ref: '#/definitions/TLSConfig'
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
    Duration:
        description: |-
            A Duration represents the elapsed time between two instants
            as an int64 nanosecond count. The representation limits the
            largest representable duration to approximately 290 years.
        format: int64
        type: integer
        x-go-package: time
    FileLoader:
        properties:
            path:
                type: string
                x-go-name: Path
        type: object
        x-go-package: github.com/go-gost/x/config
...
    HTTPLoader:
        properties:
            timeout:
                $ref: '#/definitions/Duration'
            url:
                type: string
                x-go-name: URL
        type: object
        x-go-package: github.com/go-gost/x/config
    HTTPNodeConfig:
        properties:
            auth:
                $ref: '#/definitions/AuthConfig'
            header:
                additionalProperties:
                    type: string
                type: object
                x-go-name: Header
            host:
                type: string
                x-go-name: Host
            rewrite:
                items:
                    $ref: '#/definitions/HTTPURLRewriteConfig'
                type: array
                x-go-name: Rewrite
        type: object
        x-go-package: github.com/go-gost/x/config
...
    HTTPURLRewriteConfig:
        properties:
            Match:
                type: string
            Replacement:
                type: string
        type: object
        x-go-package: github.com/go-gost/x/config
...
    HopConfig:
        properties:
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            file:
                $ref: '#/definitions/FileLoader'
            hosts:
                type: string
                x-go-name: Hosts
            http:
                $ref: '#/definitions/HTTPLoader'
            interface:
                type: string
                x-go-name: Interface
            name:
                type: string
                x-go-name: Name
            nodes:
                items:
                    $ref: '#/definitions/NodeConfig'
                type: array
                x-go-name: Nodes
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
            resolver:
                type: string
                x-go-name: Resolver
            selector:
                $ref: '#/definitions/SelectorConfig'
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    LimiterConfig:
        properties:
            file:
                $ref: '#/definitions/FileLoader'
            http:
                $ref: '#/definitions/HTTPLoader'
            limits:
                items:
                    type: string
                type: array
                x-go-name: Limits
            name:
                type: string
                x-go-name: Name
            plugin:
                $ref: '#/definitions/PluginConfig'
            redis:
                $ref: '#/definitions/RedisLoader'
            reload:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
...
    NodeConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            bypass:
                type: string
                x-go-name: Bypass
            bypasses:
                items:
                    type: string
                type: array
                x-go-name: Bypasses
            connector:
                $ref: '#/definitions/ConnectorConfig'
            dialer:
                $ref: '#/definitions/DialerConfig'
            filter:
                $ref: '#/definitions/NodeFilterConfig'
            hosts:
                type: string
                x-go-name: Hosts
            http:
                $ref: '#/definitions/HTTPNodeConfig'
            interface:
                type: string
                x-go-name: Interface
            metadata:
                additionalProperties: {}
                type: object
                x-go-name: Metadata
            name:
                type: string
                x-go-name: Name
            network:
                type: string
                x-go-name: Network
            resolver:
                type: string
                x-go-name: Resolver
            sockopts:
                $ref: '#/definitions/SockOptsConfig'
            tls:
                $ref: '#/definitions/TLSNodeConfig'
        type: object
        x-go-package: github.com/go-gost/x/config
    NodeFilterConfig:
        properties:
            host:
                type: string
                x-go-name: Host
            path:
                type: string
                x-go-name: Path
            protocol:
                type: string
                x-go-name: Protocol
        type: object
        x-go-package: github.com/go-gost/x/config
...
    PluginConfig:
        properties:
            addr:
                type: string
                x-go-name: Addr
            timeout:
                $ref: '#/definitions/Duration'
            tls:
                $ref: '#/definitions/TLSConfig'
            token:
                type: string
                x-go-name: Token
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
...
    RedisLoader:
        properties:
            addr:
                type: string
                x-go-name: Addr
            db:
                format: int64
                type: integer
                x-go-name: DB
            key:
                type: string
                x-go-name: Key
            password:
                type: string
                x-go-name: Password
            type:
                type: string
                x-go-name: Type
        type: object
        x-go-package: github.com/go-gost/x/config
...
    Response:
        properties:
            code:
                format: int64
                type: integer
                x-go-name: Code
            msg:
                type: string
                x-go-name: Msg
        type: object
        x-go-package: github.com/go-gost/x/api
...
    SelectorConfig:
        properties:
            failTimeout:
                $ref: '#/definitions/Duration'
            maxFails:
                format: int64
                type: integer
                x-go-name: MaxFails
            strategy:
                type: string
                x-go-name: Strategy
        type: object
        x-go-package: github.com/go-gost/x/config
...
    SockOptsConfig:
        properties:
            mark:
                format: int64
                type: integer
                x-go-name: Mark
        type: object
        x-go-package: github.com/go-gost/x/config
...
    TLSConfig:
        properties:
            caFile:
                type: string
                x-go-name: CAFile
            certFile:
                type: string
                x-go-name: CertFile
            commonName:
                type: string
                x-go-name: CommonName
            keyFile:
                type: string
                x-go-name: KeyFile
            options:
                $ref: '#/definitions/TLSOptions'
            organization:
                type: string
                x-go-name: Organization
            secure:
                type: boolean
                x-go-name: Secure
            serverName:
                type: string
                x-go-name: ServerName
            validity:
                $ref: '#/definitions/Duration'
        type: object
        x-go-package: github.com/go-gost/x/config
    TLSNodeConfig:
        properties:
            options:
                $ref: '#/definitions/TLSOptions'
            secure:
                type: boolean
                x-go-name: Secure
            serverName:
                type: string
                x-go-name: ServerName
        type: object
        x-go-package: github.com/go-gost/x/config
    TLSOptions:
        properties:
            cipherSuites:
                items:
                    type: string
                type: array
                x-go-name: CipherSuites
            maxVersion:
                type: string
                x-go-name: MaxVersion
            minVersion:
                type: string
                x-go-name: MinVersion
        type: object
        x-go-package: github.com/go-gost/x/config
info:
    title: Documentation of Web API.
    version: 1.0.0
paths:
...
    /config/chains:
        post:
            operationId: createChainRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ChainConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createChainResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new chain, the name of chain must be unique in chain list.
            tags:
                - Chain
    /config/chains/{chain}:
        delete:
            operationId: deleteChainRequest
            parameters:
                - in: path
                  name: chain
                  required: true
                  type: string
                  x-go-name: Chain
            responses:
                "200":
                    $ref: '#/responses/deleteChainResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete chain by name.
            tags:
                - Chain
        put:
            operationId: updateChainRequest
            parameters:
                - in: path
                  name: chain
                  required: true
                  type: string
                  x-go-name: Chain
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/ChainConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateChainResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update chain by name, the chain must already exist.
            tags:
                - Chain
    /config/climiters:
        post:
            operationId: createConnLimiterRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/LimiterConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createConnLimiterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new conn limiter, the name of limiter must be unique in limiter list.
            tags:
                - Limiter
    /config/climiters/{limiter}:
        delete:
            operationId: deleteConnLimiterRequest
            parameters:
                - in: path
                  name: limiter
                  required: true
                  type: string
                  x-go-name: Limiter
            responses:
                "200":
                    $ref: '#/responses/deleteConnLimiterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete conn limiter by name.
            tags:
                - Limiter
        put:
            operationId: updateConnLimiterRequest
            parameters:
                - in: path
                  name: limiter
                  required: true
                  type: string
                  x-go-name: Limiter
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/LimiterConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateConnLimiterResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update conn limiter by name, the limiter must already exist.
            tags:
                - Limiter
    /config/hops:
        post:
            operationId: createHopRequest
            parameters:
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/HopConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/createHopResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Create a new hop, the name of hop must be unique in hop list.
            tags:
                - Hop
    /config/hops/{hop}:
        delete:
            operationId: deleteHopRequest
            parameters:
                - in: path
                  name: hop
                  required: true
                  type: string
                  x-go-name: Hop
            responses:
                "200":
                    $ref: '#/responses/deleteHopResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Delete hop by name.
            tags:
                - Hop
        put:
            operationId: updateHopRequest
            parameters:
                - in: path
                  name: hop
                  required: true
                  type: string
                  x-go-name: Hop
                - in: body
                  name: data
                  schema:
                    $ref: '#/definitions/HopConfig'
                  x-go-name: Data
            responses:
                "200":
                    $ref: '#/responses/updateHopResponse'
            security:
                - basicAuth:
                    - '[]'
            summary: Update hop by name, the hop must already exist.
            tags:
                - Hop
...
produces:
    - application/json
responses:
...
    createChainResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    createConnLimiterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    createHopResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    deleteChainResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    deleteConnLimiterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    deleteHopResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
    updateChainResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    updateConnLimiterResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
    updateHopResponse:
        description: successful operation.
        headers:
            Data: {}
        schema:
            $ref: '#/definitions/Response'
...
schemes:
    - https
    - http
securityDefinitions:
    basicAuth:
        type: basic
swagger: "2.0"